Refactor authentication files into feature:auth module
