+++ app/build.gradle
@@ -45,16 +45,8 @@ android {
         kotlinCompilerExtensionVersion ProjectConfig.extensionVersion
     }
     packagingOptions {
-        exclude("META-INF/DEPENDENCIES")
-        exclude("META-INF/LICENSE")
-        exclude("META-INF/LICENSE.txt")
-        exclude("META-INF/license.txt")
-        exclude("META-INF/NOTICE")
-        exclude("META-INF/NOTICE.txt")
-        exclude("META-INF/notice.txt")
-        exclude("META-INF/ASL2.0")
+        exclude("/META-INF/{AL2.0,LGPL2.1}")
         exclude("META-INF/gradle/incremental.annotation.processors")
-        exclude("META-INF/*.kotlin_module")
         exclude("META-INF/*")
     }
 }
@@ -129,4 +121,5 @@ dependencies {
     implementation(project(":core:translations"))
     implementation(project(":core:util"))
     implementation(project(":data:mongo"))
+    implementation(project(":feature:auth"))
 }
