+++ app/src/main/java/com/xdiach/diarymoodapp/navigation/NavGraph.kt
@@ -9,7 +9,6 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
@@ -18,14 +17,11 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.rememberPagerState
-import com.stevdzasan.messagebar.rememberMessageBarState
-import com.stevdzasan.onetap.rememberOneTapSignInState
+import com.xdiach.auth.navigation.authenticationRoute
 import com.xdiach.translations.R
 import com.xdiach.util.model.Mood
 import com.xdiach.util.model.RequestState
 import com.xdiach.ui.components.DisplayAlertDialog
-import com.xdiach.diarymoodapp.presentation.screens.authentication.AuthenticationScreen
-import com.xdiach.diarymoodapp.presentation.screens.authentication.AuthenticationViewModel
 import com.xdiach.diarymoodapp.presentation.screens.home.HomeScreen
 import com.xdiach.diarymoodapp.presentation.screens.home.HomeViewModel
 import com.xdiach.diarymoodapp.presentation.screens.write.WriteScreen
@@ -73,66 +69,6 @@ fun SetupNavGraph(
             onBackPressed = {
                 navController.popBackStack()
             }
-        )
-    }
-}
-
-fun NavGraphBuilder.authenticationRoute(
-    navigateToHome: () -> Unit,
-    onDataLoaded: () -> Unit
-) {
-    composable(route = Screen.Authentication.route) {
-        val viewModel: AuthenticationViewModel = viewModel()
-        val authenticated by viewModel.authenticated
-        val loadingState by viewModel.loadingState
-        val oneTapAuthState = rememberOneTapSignInState()
-        val messageBarState = rememberMessageBarState()
-        val context = LocalContext.current
-
-        LaunchedEffect(key1 = Unit) {
-            onDataLoaded()
-        }
-
-        AuthenticationScreen(
-            authenticated = authenticated,
-            loadingState = loadingState,
-            oneTapAuthState = oneTapAuthState,
-            messageBarState = messageBarState,
-            onAuthClicked = {
-                oneTapAuthState.open()
-                viewModel.setLoading(true)
-            },
-            onSuccessfulFirebaseSignIn = { tokenId ->
-                viewModel.signInWithMongoAtlas(
-                    tokenId = tokenId,
-                    onSuccess = {
-                        messageBarState.addSuccess(
-                            UiText.StringResource(
-                                resId = R.string.auth_screen_signin_successful
-                            ).asString(context)
-                        )
-                        viewModel.setLoading(false)
-                    },
-                    onError = { exception ->
-                        messageBarState.addError(
-                            exception ?: Exception(
-                                UiText.StringResource(R.string.auth_screen_signin_failure)
-                                    .asString(context)
-                            )
-                        )
-                        viewModel.setLoading(false)
-                    }
-                )
-            },
-            onFailedFirebaseSignIn = { exception ->
-                messageBarState.addError(exception)
-                viewModel.setLoading(false)
-            },
-            onDialogDismissed = { message ->
-                messageBarState.addError(Exception(message))
-                viewModel.setLoading(false)
-            },
-            navigateToHome = navigateToHome
         )
     }
 }
