+++ settings.gradle
@@ -19,3 +19,4 @@ include ':core:ui'
 include ':core:translations'
 include ':core:util'
 include ':data:mongo'
+include ':feature:auth'
